% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjustment.R
\name{adj.ens}
\alias{adj.ens}
\title{adj.ens}
\usage{
adj.ens(Pf, X, mu.f, mu.a, Pa)
}
\arguments{
\item{Pf}{A cov matrix of forecast state variables.}

\item{X}{Dataframe or matrix of forecast state variables for different ensembles.}

\item{mu.f}{A vector with forecast mean estimates of state variables.}

\item{mu.a}{A vector with analysis mean estimates of state variables.}

\item{Pa}{The state estimate cov matrix of analysis.}
}
\value{
Returns a vector of adjusted analysis mean estimates of state variables.
}
\description{
This functions gives weights to different ensemble members based on their likelihood during the analysis step. Then it adjusts the analysis mean estimates of state variables based on the estimated weights.
}
\author{
Michael Dietze \email{dietze@bu.edu}, Ann Raiho and Hamze Dokoohaki
}
