% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R
\name{analysis_sda_block}
\alias{analysis_sda_block}
\title{analysis_sda_block}
\usage{
analysis_sda_block(
  settings,
  block.list.all,
  X,
  obs.mean,
  obs.cov,
  t,
  nt,
  MCMC.args,
  block.list.all.pre = NULL
)
}
\arguments{
\item{settings}{pecan standard multi-site settings list.}

\item{block.list.all}{List contains nt empty sub-elements.}

\item{X}{A matrix contains ensemble forecasts.}

\item{obs.mean}{List of dataframe of observation means, named with observation datetime.}

\item{obs.cov}{List of covariance matrices of state variables , named with observation datetime.}

\item{t}{time point.}

\item{nt}{total length of time steps.}

\item{MCMC.args}{arguments for the MCMC sampling.}

\item{block.list.all.pre}{pre-existed block.list.all object for passing the aqq and bqq to the current SDA run, the default is NULL.}
}
\value{
It returns the `build.block.xy` object and the analysis results.
}
\description{
This function provides the block-based MCMC sampling approach.
}
\details{
This function will add data and constants into each block that are needed for the MCMC sampling.
}
\author{
Dongchen Zhang
}
