% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.params.R
\name{assess.params}
\alias{assess.params}
\alias{assessParams}
\title{assess.params}
\usage{
assessParams(dat, Xt, wts = NULL, mu_f_TRUE = NULL, P_f_TRUE = NULL)
}
\arguments{
\item{dat}{MCMC output}

\item{Xt}{ensemble output matrix}

\item{wts}{ensemble weights}

\item{mu_f_TRUE}{muf before tobit2space}

\item{P_f_TRUE}{Pf before tobit2space}
}
\value{
make plots
}
\description{
Assessing parameter estimations after mapping model output to tobit space
}
\author{
Michael Dietze and Ann Raiho \email{dietze@bu.edu}
}
