% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R
\name{block.2.vector}
\alias{block.2.vector}
\title{block.2.vector}
\usage{
block.2.vector(block.list, X, H)
}
\arguments{
\item{block.list}{lists of blocks generated by the `build.block.xy` function.}

\item{X}{A matrix contains ensemble forecasts.}

\item{H}{H index created by the `construct_nimble_H` function.}
}
\value{
It returns a list of analysis results by MCMC sampling.
}
\description{
block.2.vector
}
\author{
Dongchen Zhang
}
