% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Site_Constructors.R
\name{block_matrix}
\alias{block_matrix}
\title{block_matrix}
\usage{
block_matrix(x = NULL, b = NULL, byrow = FALSE, dimnames = NULL)
}
\arguments{
\item{x}{Vector of numbers to identify each block.}

\item{b}{Numeric value for the size of the blocks within the matrix ordered depending on byrow}

\item{byrow}{ogical value. If FALSE (the default) the blocks are filled by columns, otherwise the blocks in the matrix are filled by rows.}

\item{dimnames}{Character string of name attribute for the basis of the blcok matrix. If NULL a vector of the same length of b provides the basis of row and column names.#'.}
}
\value{
Returns a matrix with block sizes determined by the b argument. Each block is filled with the same value taken from x.
}
\description{
This function is adopted from migest package.
}
\author{
Guy J. Abel
}
