% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R
\name{build.block.xy}
\alias{build.block.xy}
\title{build.block.xy}
\usage{
build.block.xy(settings, block.list.all, X, obs.mean, obs.cov, t)
}
\arguments{
\item{settings}{pecan standard multi-site settings list.}

\item{block.list.all}{List contains nt empty sub-elements.}

\item{X}{A matrix contains ensemble forecasts.}

\item{obs.mean}{List of dataframe of observation means, named with observation datetime.}

\item{obs.cov}{List of covariance matrices of state variables , named with observation datetime.}

\item{t}{time point.}
}
\value{
It returns the `build.block.xy` object with data and constants filled in.
}
\description{
This function split long vector and covariance matrix into blocks corresponding to the localization.
}
\details{
This function will add data and constants into each block that are needed for the MCMC sampling.
}
\author{
Dongchen Zhang
}
