% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_X.R
\name{build_X}
\alias{build_X}
\title{build_X}
\usage{
build_X(
  out.configs,
  settings,
  new.params,
  nens,
  read_restart_times,
  outdir = paste0(old.dir, "out/"),
  t = 1
)
}
\arguments{
\item{out.configs}{object created for build_X passed from sda.enkf_MultiSite}

\item{settings}{settings object, passed from sda.enkf_MultiSite}

\item{new.params}{object created from sda_matchparam, passed from sda.enkf_MultiSite}

\item{nens}{number of ensemble members i.e. runs}

\item{read_restart_times}{passed from sda.enkf_MultiSite}

\item{outdir}{location of previous run output folder containing .nc files}

\item{t}{Default t=1, for function to work within time loop}
}
\value{
X ready to be passed to SDA Analysis code
}
\description{
builds X matrix for SDA
}
\author{
Alexis Helgeson
}
