% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_parallel.R
\name{check.qsub.job.info}
\alias{check.qsub.job.info}
\title{sda_assemble}
\usage{
check.qsub.job.info(username)
}
\arguments{
\item{username}{character: username for the HPC.}
}
\value{
list: number of jobs and vector of job ids.
}
\description{
This function can help to assemble sda outputs (analysis and forecasts) from each job execution.
}
\author{
Dongchen Zhang.
}
