% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\name{conj_wt_wishart_sampler}
\alias{conj_wt_wishart_sampler}
\title{Weighted conjugate wishart}
\usage{
conj_wt_wishart_sampler(model, mvSaved, target, control)
}
\description{
Weighted conjugate wishart
}
