% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda.R
\name{Construc_H}
\alias{Construc_H}
\title{Construc_H}
\usage{
Construc_H(choose, Y, X)
}
\arguments{
\item{choose}{}

\item{Y}{vector of observations}

\item{X}{Dataframe or matrix of forecast state variables for different ensembles.}
}
\value{
This returns a mtrix specifying which observation go with which state variables.
}
\description{
This function creates a mtrix mapping obsereved data to their forecast state variable.
}
\author{
Hamze Dokoohaki
}
