% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Site_Constructors.R
\name{Construct.H.multisite}
\alias{Construct.H.multisite}
\title{Construct.H.multisite}
\usage{
Construct.H.multisite(site.ids, var.names, obs.t.mean)
}
\arguments{
\item{site.ids}{a vector name of site ids}

\item{var.names}{vector names of state variable names}

\item{obs.t.mean}{list of vector of means for the time t for different sites.}
}
\value{
Returns a matrix with block sizes determined by the b argument. Each block is filled with the same value taken from x.
}
\description{
This function is makes the blocked mapping function.
}
\author{
Hamze
}
