% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Site_Constructors.R
\name{Construct.R}
\alias{Construct.R}
\title{Construct.R}
\usage{
Construct.R(site.ids, var.names, obs.t.mean, obs.t.cov)
}
\arguments{
\item{site.ids}{a vector name of site ids}

\item{var.names}{vector names of state variable names}

\item{obs.t.mean}{list of vector of means for the time t for different sites.}

\item{obs.t.cov}{list of list of cov for the time for different sites.}
}
\value{
This function returns a list with Y and R ready to be sent to the analysis functions.
}
\description{
Make sure that both lists are named with siteids.
}
\author{
Hamze Dokoohaki
}
