% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Site_Constructors.R
\name{construct_nimble_H}
\alias{construct_nimble_H}
\title{construct_nimble_H}
\usage{
construct_nimble_H(site.ids, var.names, obs.t, pft.path = NULL, by = "single")
}
\arguments{
\item{site.ids}{a vector name of site ids}

\item{var.names}{vector names of state variable names}

\item{obs.t}{list of vector of means for the time t for different sites.}

\item{pft.path}{physical path to the pft.csv file.}

\item{by}{criteria, it supports by variable, site, pft, all, and single Q.}
}
\value{
Returns one vector containing index for which Q to be estimated for which variable,
and the other vector gives which state variable has which observation (= element.W.Data).
}
\description{
This function is an upgrade to the Construct.H.multisite function which provides the index by different criteria.
}
\author{
Dongchen Zhang
}
