% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Site_Constructors.R
\name{Contruct.Pf}
\alias{Contruct.Pf}
\title{Contruct.Pf}
\usage{
Contruct.Pf(site.ids, var.names, X)
}
\arguments{
\item{site.ids}{a vector name of site ids}

\item{var.names}{vector names of state variable names}

\item{X}{a matrix of state variables. In this matrix rows represent ensembles, while columns show the variables for different sites.}
}
\value{
It returns the var-cov matrix of state variables at multiple sites.
}
\description{
The argument X needs to have an atrribute pointing the state variables to their corresponding site. This atrribute needs to be called `Site`.
At the moment, the cov between state variables at block defining the cov between two sites are assumed zero.
}
\author{
Hamze Dokoohaki
}
