% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_bias_correction.R
\name{debias.average}
\alias{debias.average}
\title{debias.average}
\usage{
debias.average(residuals)
}
\arguments{
\item{residuals}{vector: predicted residuals across products}
}
\value{
numeric: an averaged residual.
}
\description{
This function helps to average the predicted residuals based on certain criteria.
The default function is averaging.
TODO: we will be averaging residuals based on ML precision for each residual prediction.
}
\author{
Dongchen Zhang
}
