% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_bias_correction.R
\name{debias.extract.cov}
\alias{debias.extract.cov}
\title{debias.extract.cov}
\usage{
debias.extract.cov(cov.dir, date, site.info)
}
\arguments{
\item{cov.dir}{character: physical path to the directory contains the time series covariate maps.}

\item{date}{numeric or character: date that used to find the corresponding covariate file.}

\item{site.info}{data.frame: data.frame that contains longitude and latitude in its first and second column.}
}
\value{
data.frame: a data frame containing M (rows) of locations across N (columns) variables.
}
\description{
This function helps to extract the covariates based on 
date and coordinates from the settings object.
}
\author{
Dongchen Zhang
}
