% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_bias_correction.R
\name{debias.map.products}
\alias{debias.map.products}
\title{debias.map.products}
\usage{
debias.map.products(obs.mean.t, obs_prep)
}
\arguments{
\item{obs.mean.t}{list: lists of sites that contain observations for each variable.}

\item{obs_prep}{list: lists contains basic information for each product of each observation.}
}
\value{
matrix: a matrix containing M (rows) of locations*variables across N (columns) products.
}
\description{
This function helps to create the mapping operator that maps between product 
names and variable names by locations.
}
\author{
Dongchen Zhang
}
