% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_bias_correction.R
\name{debias.ML}
\alias{debias.ML}
\title{sda.bias.correction}
\usage{
debias.ML(pred.name, cov.names, dat.train, dat.pred, var.name, py.init)
}
\arguments{
\item{pred.name}{character: the name for the predictive variable.}

\item{cov.names}{character: the name for the covariates.}

\item{dat.train}{data.frame: data frame containing associated covariates and 
predictive variable for ML training.}

\item{dat.pred}{data.frame: data frame containing associated covariates for prediction.}

\item{var.name}{character: variable name to be predicted.}

\item{py.init}{R function: R function to initialize the python functions. Default is NULL.
the default random forest will be used if `py.init` is NULL.}
}
\value{
list: the ML predicted residuals and other ML outputs.
}
\description{
This function helps to correct the forecasts' biases based on 
ML (random forest) training on the previous time points.
}
\author{
Dongchen Zhang
}
