% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_bias_correction.R
\name{debias.res.lag.calc}
\alias{debias.res.lag.calc}
\title{debias.res.lag.calc}
\usage{
debias.res.lag.calc(res.ts)
}
\arguments{
\item{res.ts}{list: lists of residuals at each time point.}
}
\value{
list: lists of lagged residuals between time points.
}
\description{
This function helps to calculate the lagged residual error time series.
}
\author{
Dongchen Zhang
}
