% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_bias_correction.R
\name{debias.residual.calc}
\alias{debias.residual.calc}
\title{debias.residual.calc}
\usage{
debias.residual.calc(obs.mean, all.X, t, var.name)
}
\arguments{
\item{obs.mean}{List: lists of date times named by time points, which contains lists of sites named by site ids, 
which contains observation means for each state variables of each site for each time point.}

\item{all.X}{list: lists of data frame of model forecast from the beginning to the current time points 
that has n (ensemble size) rows and n.var (number of variables) times n.site (number of locations) columns.
(e.g., 100 ensembles, 4 variables, and 8,000 locations will end up with data.frame of 100 rows and 32,000 columns).}

\item{t}{numeric: the current number of time points (e.g., t=1 for the beginning time point).}

\item{var.name}{character: variable name to be predicted.}
}
\value{
list: lists of residuals between forecasts and observations across at time t.
}
\description{
This function helps to calculate the residual error for a certain time point and variable.
}
\author{
Dongchen Zhang
}
