% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_bias_correction.R
\name{debias.train.accuracy}
\alias{debias.train.accuracy}
\title{debias.train.accuracy}
\usage{
debias.train.accuracy(
  pred.name,
  cov.names,
  dat.train,
  var.name,
  py.init,
  ratio = 0.8
)
}
\arguments{
\item{pred.name}{character: the name for the predictive variable.}

\item{cov.names}{character: the name for the covariates.}

\item{dat.train}{data.frame: data frame containing associated covariates and 
predictive variable for ML training.}

\item{var.name}{character: variable name to be predicted.}

\item{py.init}{R function: R function to initialize the python functions. Default is NULL.
the default random forest will be used if `py.init` is NULL.}

\item{ratio}{numeric from 0 to 1: define the ratio of samples used for training.
The rest samples will be used for calculating the out-of-sample accuracy.
Default is 0.8.}
}
\value{
list: the variable importance of the ML and RMSE of the out-of-sample predictions.
}
\description{
This function helps to calculate the out-of-sample accuracy of residual predictions.
}
\author{
Dongchen Zhang
}
