% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance_DA.R
\name{disturbance_DA}
\alias{disturbance_DA}
\title{disturbance data assimilation.}
\usage{
disturbance_DA(
  X,
  Y,
  R,
  H,
  aqq,
  bqq,
  disturbance,
  mu0 = rep(80, 80),
  V0 = diag(rep(0.25, 2))
)
}
\arguments{
\item{X}{numeric matrix (dim = M*N, where M is the ensemble number while N is the number of variables): model forecasts for each variable of each ensemble member.}

\item{Y}{numeric vector (dim = L, the number of observations): vector stores the available observations for the corresponding variables.}

\item{R}{numeric matrix (dim = L*L): the covariance matrix of the observations.}

\item{H}{integral vector (dim = L): mapping operator (Y -> X. e.g., c(1,3,4) represents the fact that we have three observations covered the 1st, 2nd, and 4th state variables).}

\item{aqq}{numeric vector (dim = N): the vector of the `shape` parameter for the gamma distribution for each state variable.}

\item{bqq}{numeric vector (dim = N): the vector of the `rate` parameter for the gamma distribution for each state variable.}

\item{disturbance}{logical (dim = 1): if the current site experienced the disturbance (1) or not (0).}

\item{mu0}{numeric vector (dim = B where B is the number of biomass variables): percent decreases of biomass to be used for the disturbance model}

\item{V0}{numeric matrix (dim B*B): covariance to be used for the disturbance model}
}
\value{
returns a vector of mua (dim = N, vector of means for each updated state variable) and pa (dim = N*N, covariance matrix for the updated state variables).
}
\description{
This function provides semi-analytical updates for the disturbance probability.
}
