% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_NA_downscale.R
\name{downscale.qsub.main}
\alias{downscale.qsub.main}
\title{downscale.qsub.main}
\usage{
downscale.qsub.main(folder.path)
}
\arguments{
\item{folder.path}{Character: physical path to which the job file is located.}
}
\description{
This qsub function helps to run the submitted qsub jobs for running the downscale.rf.main function.
}
\author{
Dongchen Zhang
}
