% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_parallel_downscale.R
\name{downscale_main}
\alias{downscale_main}
\title{downscale_main}
\usage{
downscale_main(
  settings,
  analysis,
  covariates.dir,
  time,
  variable,
  outdir,
  base.map.dir,
  method = "randomForest",
  cores = parallel::detectCores()
)
}
\arguments{
\item{settings}{character: physical path that points to the pecan settings XML file.}

\item{analysis}{numeric: data frame (rows: ensemble member; columns: site*state_variables) of updated ensemble analysis results from the `sda_enkf` function.}

\item{covariates.dir}{character: path to the exported covariates GeoTIFF file.}

\item{time}{character: the time tag used to differentiate the outputs from others.}

\item{variable}{character: name of state variable. It should match up with the column names of the analysis data frame.}

\item{outdir}{character: the output directory where the downscaled maps will be stored.}

\item{base.map.dir}{character: path to the GeoTIFF file within which the extents and CRS will be used to generate the ensemble maps.}

\item{method}{character: machine learning method, default is randomForest (currently support randomForest and xgboost).}

\item{cores}{numeric: how many CPus to be used in the calculation, the default is the total CPU number you have.}
}
\value{
paths to the ensemble downscaled maps.
}
\description{
This is the main function to execute the machine learning training and prediction. 
Note it will be deployed by each node you requested if the qsub feature is enabled below.
}
\author{
Dongchen Zhang
}
