% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_parallel_downscale.R
\name{downscale_qsub_main}
\alias{downscale_qsub_main}
\title{downscale_qsub_main}
\usage{
downscale_qsub_main(folder.path)
}
\arguments{
\item{folder.path}{Character: physical path to which the job file is located.}
}
\description{
This qsub function helps to run the submitted qsub jobs for running the downscale_main function.
}
\author{
Dongchen Zhang
}
