% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_multiSite.R
\name{EnKF.MultiSite}
\alias{EnKF.MultiSite}
\title{EnKF.MultiSite}
\usage{
EnKF.MultiSite(settings, Forecast, Observed, H, extraArg = NULL, ...)
}
\arguments{
\item{settings}{pecan standard settings list.}

\item{Forecast}{A list containing the forecasts variables including Q (process variance) and X (a dataframe of forecasts state variables for different ensemble)}

\item{Observed}{A list containing the observed variables including R (cov of observed state variables) and Y (vector of estimated mean of observed state variables)}

\item{H}{is a matrix of 1's and 0's specifying which observations go with which state variables.}

\item{extraArg}{This argument is NOT used inside this function but it is a list containing aqq, bqq and t. The aqq and bqq are shape parameters estimated over time for the proccess covariance and t gives the time in terms of index of obs.list.}

\item{...}{Extra argument sent to the analysis function.}
}
\value{
It returns a list with estimated mean and cov matrix of forecast state variables as well as mean and cov estimated as a result of assimilation/analysis .
}
\description{
Given the Forecast and Observed this function performs the Ensemble Kalamn Filter.
}
\details{
This function is different than `EnKF` function in terms of how it creates the Pf matrix.
}
\author{
Michael Dietze \email{dietze@bu.edu}, Ann Raiho and Hamze Dokoohaki
}
