% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_downscale.R
\name{ensemble_downscale}
\alias{ensemble_downscale}
\title{Ensemble Downscale}
\usage{
ensemble_downscale(ensemble_data, site_coords, covariates)
}
\arguments{
\item{ensemble_data}{EFI standard tibble or data.frame. Contains carbon data for downscaling.}

\item{site_coords}{data.frame, tibble, or sf object. Design points. If not sf object, must have
'lon' and 'lat' columns. Must have a unique `site_id` field.}

\item{covariates}{table containing numeric predictors to be used in downscaling.
Must have unique identifier 'site_id' field and predictor attributes}
}
\value{
A list containing the model, predictions for all values of covariates as well as test data and test predictions for downstream
statistics.
}
\description{
Ensemble Downscale
}
\details{
This function will downscale forecast data to unmodeled locations using covariates and site locations
}
\author{
Joshua Ploshay, Sambhav Dixit, David LeBauer
}
