% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\docType{data}
\name{GEF.MultiSite.Nimble}
\alias{GEF.MultiSite.Nimble}
\title{multisite TWEnF}
\format{An object of class \code{{} of length 9.}
\usage{
GEF.MultiSite.Nimble
}
\description{
multisite TWEnF
}
\keyword{datasets}
