% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_plotting.R
\name{generate_colors_sda}
\alias{generate_colors_sda}
\title{generate_colors_sda}
\usage{
generate_colors_sda()
}
\description{
This function generates a series of colors in its parents enviroment. This is mainly used in assim.sequential package.
}
\author{
Ann Raiho
}
