% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensemble_weights.R
\name{get_ensemble_weights}
\alias{get_ensemble_weights}
\title{get_ensemble_weights}
\usage{
get_ensemble_weights(settings)
}
\arguments{
\item{settings}{PEcAn settings object}
}
\value{
NONE
}
\description{
Creates file of ensemble weights in format needed for SDA
}
\author{
Ann Raiho \email{ann.raiho@gmail.com}
}
