% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{GrabFillMatrix}
\alias{GrabFillMatrix}
\title{GrabFillMatrix}
\usage{
GrabFillMatrix(M, ind, M1 = NULL)
}
\arguments{
\item{M}{source matrix that will be either subtracted or filled in.}

\item{ind}{vector of index that of where to be subtracted or filled in.}

\item{M1}{additional matrix used to fill in the source matrix, the default it NULL.}
}
\description{
GrabFillMatrix
}
\details{
This function helps subtract or fill in a matrix given the index.
}
\author{
Dongchen Zhang
}
