% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hop_test.R
\name{hop_test}
\alias{hop_test}
\title{hop_test}
\usage{
hop_test(settings, ens.runid = NULL)
}
\arguments{
\item{settings}{SDA PEcAn settings object}
}
\value{
NONE
}
\description{
Hop test. This script tests that the model successfully reads it's own restart and can restart without loss of information.
}
\author{
Ann Raiho \email{araiho@nd.edu}
}
