% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_plotting.R
\name{interactive.plotting.sda}
\alias{interactive.plotting.sda}
\alias{postana.timeser.plotting.sda}
\alias{postana.bias.plotting.sda}
\alias{postana.bias.plotting.sda.corr}
\alias{post.analysis.ggplot}
\alias{post.analysis.ggplot.violin}
\alias{post.analysis.multisite.ggplot}
\alias{SDA_timeseries_plot}
\title{Internal functions for plotting SDA outputs. Interactive, post analysis time-series and bias plots in base plotting system and ggplot}
\usage{
interactive.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.timeser.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.bias.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.bias.plotting.sda.corr(t, obs.times, X, aqq, bqq)

post.analysis.ggplot(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS,
  plot.title = NULL
)

post.analysis.ggplot.violin(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS,
  plot.title = NULL
)

post.analysis.multisite.ggplot(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  FORECAST,
  ANALYSIS,
  plot.title = NULL,
  facetg = FALSE,
  readsFF = NULL,
  Add_Map = FALSE
)

SDA_timeseries_plot(
  ANALYSIS,
  FORECAST,
  obs.mean,
  obs.cov,
  outdir,
  by = "site",
  CI = c(0.025, 0.975),
  unit = list(AbvGrndWood = "Mg/ha", LAI = "m2/m2", SoilMoistFrac = "", TotSoilCarb =
    "kg/m2")
)
}
\arguments{
\item{settings}{pecan standard settings list.}

\item{t}{current time - int number giving the position of the current time in obs.time.}

\item{obs.mean}{list of vectors of the means of observed data named by the measured date.}

\item{obs.cov}{list of cov matrices of the observed data named by the measured date.}

\item{obs}{list containing the mean and cov object}

\item{X}{dataframe of state variables for each ensemble}

\item{FORECAST}{dataframe of state variables for each ensemble}

\item{ANALYSIS}{vector of mean of state variable after analysis}

\item{plot.title}{character giving the title for post visualization ggplots}

\item{Add_Map}{Bool variable decide if we want to export the GIS map of Ecoregion.}

\item{obs.time}{vector of dates of measurements}
}
\description{
Internal functions for plotting SDA outputs. Interactive, post analysis time-series and bias plots in base plotting system and ggplot
}
