% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_paleon_sda.R
\name{load_data_paleon_sda}
\alias{load_data_paleon_sda}
\title{load_data_paleon_sda}
\usage{
load_data_paleon_sda(settings)
}
\arguments{
\item{settings}{PEcAn SDA settings object}
}
\value{
obs.mean and obs.cov for sda.enkf
}
\description{
Load data function for paleon SDA data products
}
\author{
Ann Raiho \email{araiho@nd.edu}
}
