% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_MultiSite.R
\name{loadRData}
\alias{loadRData}
\title{loadRData}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{path to the Rdata.
Enable the feature of passing loaded object into new variable.}
}
\description{
loadRData
}
