% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Localization.R
\name{Local.support}
\alias{Local.support}
\title{Local.support}
\usage{
Local.support(Pf)
}
\arguments{
\item{Pf}{Forecast error coveriance matrix}
}
\value{
It returns a localized covariance matrix by taking a Schur product between Pf and a corrolation function
}
\description{
Local.support
}
\author{
Hamze Dokoohaki
}
