% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operation.R
\name{matrix_network}
\alias{matrix_network}
\title{matrix_network}
\usage{
matrix_network(mat)
}
\arguments{
\item{mat}{a boolean matrix representing the interactions between any sites.}
}
\value{
It returns lists of index representing each network.
}
\description{
matrix_network
}
\author{
Dongchen Zhang
}
