% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R
\name{MCMC_block_function}
\alias{MCMC_block_function}
\title{MCMC_block_function}
\usage{
MCMC_block_function(block)
}
\arguments{
\item{block}{each block within the `block.list` lists.}
}
\value{
It returns the `block` object with analysis results filled in.
}
\description{
MCMC_block_function
}
\author{
Dongchen Zhang
}
