% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_multiSite.R
\name{MCMC_function}
\alias{MCMC_function}
\title{MCMC_function}
\usage{
MCMC_function(data)
}
\arguments{
\item{data}{list containing everything needed for the MCMC sampling.}
}
\description{
MCMC_function
}
\details{
This function replace the previous code where implenmented the MCMC sampling part, which allows the MCMC sampling of multiple chains under parallel mode.
}
\author{
Michael Dietze \email{dietze@bu.edu}, Ann Raiho, Hamze Dokoohaki, and Dongchen Zhang.
}
