% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R,
%   R/Analysis_sda_block_old.R
\name{MCMC_Init}
\alias{MCMC_Init}
\title{MCMC_Init}
\usage{
MCMC_Init(block.list, X)

MCMC_Init(block.list, X)
}
\arguments{
\item{block.list}{lists of blocks generated by the `build.block.xy` function.}

\item{X}{A matrix contains ensemble forecasts.}
}
\value{
It returns the `block.list` object with initial conditions filled in.

It returns the `block.list` object with initial conditions filled in.
}
\description{
MCMC_Init

MCMC_Init
}
\details{
This function helps create initial conditions for the MCMC sampling.

This function helps create initial conditions for the MCMC sampling.
}
\author{
Dongchen Zhang
}
