% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metSplit.R
\name{metSplit}
\alias{metSplit}
\title{metSplit}
\usage{
metSplit(
  conf.settings,
  inputs,
  settings,
  model,
  no_split = FALSE,
  obs.times,
  t,
  nens,
  restart_flag = FALSE,
  my.split_inputs
)
}
\arguments{
\item{conf.settings}{listed multisite settings object generated by sda.enkf_MultiSite}

\item{inputs}{listed object containing met ensemble members generated by input.ens.gen}

\item{settings}{settings object passed to sda.enkf_MultiSite}

\item{model}{model name ex. SIPNET}

\item{no_split}{TRUE/FALSE if model requires met split}

\item{obs.times}{matrix of dates used for assimilation}

\item{t}{number of dates in obs.times}

\item{nens}{number of ensemble members, taken from settings object}

\item{restart_flag}{TRUE/FALSE taken from restart arguement}

\item{my.split_inputs}{generated by sda.enkf_MultiSite ex. split_inputs.SIPNET}
}
\value{
input.split object with split met filepaths
}
\description{
metSplit
}
\author{
Alexis Helgeson
}
