% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_function_hrly.R
\name{NA_downscale_hrly}
\alias{NA_downscale_hrly}
\title{North America Downscale Function}
\usage{
NA_downscale_hrly(nc_data, coords, date, covariates)
}
\arguments{
\item{nc_data}{In quotes, file path for .nc containing ensemble data.}

\item{coords}{In quotes, file path for .csv file containing the site coordinates, columns named "lon" and "lat".}

\item{date}{In quotes, format is yyyy-mm-dd hh:mm:ss EST. Restricted to time within file supplied to 'data' (hours since 1986-01-01T00:00).}

\item{covariates}{SpatRaster stack, used as predictors in randomForest. Layers within stack should be named. Recommended that this stack be generated using 'covariates' instructions in assim.sequential/inst folder}
}
\value{
It returns the `downscale_output` list containing lists for the training and testing data sets, models, and predicted maps for each ensemble member.
}
\description{
This function uses the randomForest model.
}
\details{
This function will downscale forecast data (hourly) to unmodeled locations using covariates and site locations
}
\author{
Harunobu Ishii
}
