% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.functions.R
\name{obs_timestep2timepoint}
\alias{obs_timestep2timepoint}
\title{convert from timestep to actual time points.
supports year, month, week, and day as time unit.}
\usage{
obs_timestep2timepoint(start.date, end.date, timestep)
}
\arguments{
\item{start.date}{start date when the first observation was taken.}

\item{end.date}{end date when the last observation was taken.}

\item{timestep}{a list includes time unit and number of time unit per timestep.}
}
\value{
timepoints from start to end date given the number of time unit per timestep.
}
\description{
convert from timestep to actual time points.
supports year, month, week, and day as time unit.
}
\author{
Dongchen Zhang
}
