% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.functions.R
\name{outlier_detector_boxplot}
\alias{outlier_detector_boxplot}
\title{Replace outliers using boxplot statistics}
\usage{
outlier_detector_boxplot(X)
}
\arguments{
\item{X}{A list of dataframes}
}
\value{
A list the same dimension as X, with each column of each dataframe
  modified by replacing outlier points with the column median
}
\description{
This function performs a simple outlier replacement on all the columns of dataframes inside a list
}
