% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.batch.functions.R
\name{parallel.job.execution}
\alias{parallel.job.execution}
\title{parallel.job.execution}
\usage{
parallel.job.execution(folder.path, cores)
}
\arguments{
\item{folder.path}{Character: physical path to which the job file is located.}

\item{cores}{numeric: number of requested CPUs.}
}
\description{
This function realizes the parallel model execution.
}
\author{
Dongchen Zhang.
}
