% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.qsub.functions.R
\name{parallel.read.sda}
\alias{parallel.read.sda}
\title{parallel.read.sda}
\usage{
parallel.read.sda(
  settings,
  my.read_restart,
  outdir,
  out.configs,
  stop.time,
  var.names,
  new.params
)
}
\arguments{
\item{settings}{PEcAn settings object.}

\item{my.read_restart}{Character: Function to be called for reading sda outputs.}

\item{outdir}{Character: physical path to which model outputs are stored.}

\item{out.configs}{Lists: outputs returned by `parallel.write.configs` function.}

\item{stop.time}{Character: stop time for reading NC files.}

\item{var.names}{Character: names of variables to be read.}

\item{new.params}{Lists: outputs returned by `sda_matchparam` function.}
}
\description{
This function provides an option for reading sda outputs using cluster.
}
\author{
Dongchen Zhang.
}
