% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_NA_downscale.R
\name{parallel.rf.train}
\alias{parallel.rf.train}
\title{parallel.rf.train}
\usage{
parallel.rf.train(full_data, cores = parallel::detectCores())
}
\arguments{
\item{full_data}{numeric: the matrix generated using the `prepare.train.dat` function.}

\item{cores}{numeric: how many CPus to be used in the calculation, the default is the total CPU number you have.}
}
\value{
list of trained models across ensemble members.
}
\description{
This function helps to train the ML model across ensemble members in parallel.
}
\author{
Dongchen Zhang
}
