% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.qsub.functions.R
\name{parallel.split.met}
\alias{parallel.split.met}
\title{parallel.split.met}
\usage{
parallel.split.met(
  settings,
  my.split_inputs,
  conf.settings.before.split,
  start.time,
  stop.time,
  inputs
)
}
\arguments{
\item{settings}{PEcAn settings object.}

\item{my.split_inputs}{Character: Function to be called for splitting met.}

\item{conf.settings.before.split}{Lists: includes lists of configs for each site.}

\item{start.time}{Character: start time for cut met files.}

\item{stop.time}{Character: stop time for cut met files.}

\item{inputs}{List: physical paths to met files for each site.}
}
\description{
This function provides an option for splitting met files using cluster.
}
\author{
Dongchen Zhang.
}
