% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.qsub.functions.R
\name{parallel.write.configs}
\alias{parallel.write.configs}
\title{parallel.write.configs}
\usage{
parallel.write.configs(settings, ensemble.samples, restart.list)
}
\arguments{
\item{settings}{PEcAn settings object.}

\item{ensemble.samples}{Lists: Sampled parameters for each plant functional types.}

\item{restart.list}{Lists: restart lists of each site from previous runs.}
}
\description{
This function provides an option for submitting `write.configs` jobs to cluster.
}
\author{
Dongchen Zhang.
}
