% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_parallel_downscale.R
\name{parallel_train}
\alias{parallel_train}
\title{parallel_train}
\usage{
parallel_train(
  full_data,
  method = "randomForest",
  cores = parallel::detectCores()
)
}
\arguments{
\item{full_data}{numeric: the matrix generated using the `prepare_train_dat` function.}

\item{method}{character: machine learning method (currently support randomForest and xgboost).}

\item{cores}{numeric: how many CPus to be used in the calculation, the default is the total CPU number you have.}
}
\value{
list of trained models across ensemble members.
}
\description{
This function helps to train the ML model across ensemble members in parallel.
}
\author{
Dongchen Zhang
}
