% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_NA_downscale.R
\name{pecan.settings.2.pts}
\alias{pecan.settings.2.pts}
\title{pecan.settings.2.pts}
\usage{
pecan.settings.2.pts(settings)
}
\arguments{
\item{settings}{PEcAn settings: either a character that points to the settings or the actual settings object will be accepted.}
}
\value{
terra spatial points object.
}
\description{
convert settings to geospatial points in terra.
}
\author{
Dongchen Zhang
}
