% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_parallel_downscale.R
\name{pecan_settings_2_pts}
\alias{pecan_settings_2_pts}
\title{pecan_settings_2_pts}
\usage{
pecan_settings_2_pts(settings)
}
\arguments{
\item{settings}{PEcAn settings: either a character that points to the settings or shape file or the actual pecan settings object will be accepted.}
}
\value{
terra spatial points object.
}
\description{
convert settings to geospatial points in terra.
}
\author{
Dongchen Zhang
}
