% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prep_OBS_SDA.R
\name{Prep_OBS_SDA}
\alias{Prep_OBS_SDA}
\title{SDA observation preparation function for LAI and AGB}
\usage{
Prep_OBS_SDA(
  multi.settings,
  start_year,
  end_year,
  out_dir,
  AGB_dir = "/projectnb/dietzelab/hamzed/LandTrendr/LandTrendr_AGB_data/"
)
}
\arguments{
\item{multi.settings}{settings objects that contains multiple sites info}

\item{start_year}{starting year}

\item{end_year}{ending year}

\item{out_dir}{output dir}

\item{AGB_dir}{AGB data dir}
}
\value{

}
\description{
SDA observation preparation function for LAI and AGB
}
