% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_NA_downscale.R
\name{prepare.train.dat}
\alias{prepare.train.dat}
\title{prepare.train.dat}
\usage{
prepare.train.dat(settings, analysis, covariates.dir, variable)
}
\arguments{
\item{settings}{character: physical path that points to the pecan settings XML file.}

\item{analysis}{numeric: data frame (rows: ensemble member; columns: site*state_variables) of updated ensemble analysis results from the `sda_enkf` function.}

\item{covariates.dir}{character: path to the exported covariates GeoTIFF file.}

\item{variable}{character: name of state variable. It should match up with the column names of the analysis data frame.}
}
\value{
matrix within which the first sets of columns contain values of state variables for each ensemble mebers of every site, and the rest columns contain the corresponding covariates.
}
\description{
This function helps to create the training dataset of specific variable type and locations for downscaling.
TODO: There will be a ratio argument between training and testing samples to testify the ML regression accuracy.
}
\author{
Dongchen Zhang
}
