% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_parallel_downscale.R
\name{prepare_train_dat}
\alias{prepare_train_dat}
\title{prepare_train_dat}
\usage{
prepare_train_dat(pts, analysis, covariates.dir, variable)
}
\arguments{
\item{pts}{spatialpoints: spatial points returned by `terra::vectors` function.}

\item{analysis}{numeric: data frame (rows: ensemble member; columns: site*state_variables) of updated ensemble analysis results from the `sda_enkf` function.}

\item{covariates.dir}{character: path to the exported covariates GeoTIFF file.}

\item{variable}{character: name of state variable. It should match up with the column names of the analysis data frame.}
}
\value{
matrix (num.sites, num.variables * num.ensemble + num.covariates) within which the first sets of columns contain values of state variables for each ensemble member of every site, and the rest columns contain the corresponding covariates.
}
\description{
This function helps to create the training dataset of specific variable type and locations for downscaling.
TODO: Add a ratio argument (training sample size/total sample size) so that we could calculate the out-of-sample accuracy.
}
\author{
Dongchen Zhang
}
