% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R
\name{qsub_analysis}
\alias{qsub_analysis}
\title{qsub_analysis}
\usage{
qsub_analysis(folder.path)
}
\arguments{
\item{folder.path}{character: path where the `block.Rdata` file is stored.}
}
\description{
This function can help to execute `foreach` parallel MCMC sampling given generated MCMC configuration lists.
}
