% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.qsub.functions.R
\name{qsub_analysis_submission}
\alias{qsub_analysis_submission}
\title{qsub_analysis_submission}
\usage{
qsub_analysis_submission(settings, block.list)
}
\arguments{
\item{settings}{PEcAn settings object.}

\item{block.list}{list: MCMC configuration lists for the block SDA analysis.}
}
\description{
This function provides means to split large SDA analysis (MCMC) runs into separate `qsub` jobs.
Including job creation, submission, and assemble.
}
