% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_parallel.R
\name{qsub_sda_batch}
\alias{qsub_sda_batch}
\title{qsub_sda_batch}
\usage{
qsub_sda_batch(folder.path)
}
\arguments{
\item{folder.path}{character: path where the `configs.rds` file is stored.}
}
\description{
This function can help to execute sda function.
}
\author{
Dongchen Zhang.
}
