% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remote_helpers.R
\name{Remote_Sync_launcher}
\alias{Remote_Sync_launcher}
\title{Remote_Sync_launcher}
\usage{
Remote_Sync_launcher(settingPath, remote.path, PID)
}
\arguments{
\item{settingPath}{Path to your local setting .}

\item{remote.path}{Path generated by SDA_remote_launcher which shows the path to your remote SDA run.}

\item{PID}{PID generated by SDA_remote_launcher which shows the active PID running your SDA job.}
}
\description{
Remote_Sync_launcher
}
