% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.functions.R
\name{rescaling_stateVars}
\alias{rescaling_stateVars}
\title{rescaling_stateVars}
\usage{
rescaling_stateVars(settings, X, multiply = TRUE)
}
\arguments{
\item{settings}{pecan xml settings where state variables have the scaling_factor tag}

\item{X}{Any Matrix with column names as variable names}

\item{multiply}{TRUE = multiplication, FALSE = division}
}
\value{
rescaled Matrix
}
\description{
This function uses a set of scaling factors defined in the pecan XML to scale a given matrix
}
