% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\name{rwtmnorm}
\alias{rwtmnorm}
\title{random weighted multivariate normal}
\usage{
rwtmnorm(n, mean, prec, wt)
}
\arguments{
\item{n}{sample size}

\item{mean}{mean}

\item{prec}{precision}

\item{wt}{weight}
}
\description{
random weighted multivariate normal
}
