% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_filtering_helpers.R
\name{sample_met}
\alias{sample_met}
\title{Sample meteorological ensembles}
\usage{
sample_met(settings, nens = 1)
}
\arguments{
\item{settings}{PEcAn settings list}

\item{nens}{number of ensemble members to be sampled}
}
\description{
Sample meteorological ensembles
}
