% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\name{sampler_toggle}
\alias{sampler_toggle}
\title{sampler toggling}
\usage{
sampler_toggle(model, mvSaved, target, control)
}
\arguments{
\item{model}{model}

\item{mvSaved}{copied to}

\item{target}{thing being targetted}

\item{control}{unused}
}
\description{
sampler toggling
}
