% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf.R
\name{sda.enkf}
\alias{sda.enkf}
\title{sda.enkf}
\usage{
sda.enkf(settings, obs.mean, obs.cov, IC = NULL, Q = NULL,
  adjustment = TRUE)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{obs.mean}{list of observations of the means of state variable (time X nstate)}

\item{obs.cov}{list of observations of covariance matrices of state variables (time X nstate X nstate)}

\item{IC}{initial conditions}

\item{Q}{process covariance matrix given if there is no data to estimate it}

\item{adjustment}{flag for using ensemble adjustment filter or not}
}
\value{
NONE
}
\description{
State Variable Data Assimilation: Ensemble Kalman Filter
}
\author{
Michael Dietze and Ann Raiho \email{dietze@bu.edu}
}
