% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_NorthAmerica.R
\name{sda.enkf_NorthAmerica}
\alias{sda.enkf_NorthAmerica}
\title{sda.enkf_NorthAmerica}
\usage{
sda.enkf_NorthAmerica(
  settings,
  obs.mean,
  obs.cov,
  Q = NULL,
  pre_enkf_params = NULL,
  ensemble.samples = NULL,
  control = list(TimeseriesPlot = FALSE, OutlierDetection = FALSE, send_email = NULL,
    keepNC = TRUE, forceRun = TRUE, MCMC.args = NULL),
  ...
)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{obs.mean}{Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation means for each state variables of each site for each time point.}

\item{obs.cov}{Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation covariances for all state variables of each site for each time point.}

\item{Q}{Process covariance matrix given if there is no data to estimate it.}

\item{pre_enkf_params}{Used for passing pre-existing time-series of process error into the current SDA runs to ignore the impact by the differences between process errors.}

\item{ensemble.samples}{Pass ensemble.samples from outside to avoid GitHub check issues.}

\item{control}{List of flags controlling the behavior of the SDA. 
`TimeseriesPlot` for post analysis examination; 
`OutlierDetection` decide if we want to execute the outlier detection each time after the model forecasting;
`send_email` contains lists for sending email to report the SDA progress;
`keepNC` decide if we want to keep the NetCDF files inside the out directory;
`forceRun` decide if we want to proceed the Bayesian MCMC sampling without observations;
`MCMC.args` include lists for controling the MCMC sampling process (iteration, nchains, burnin, and nthin.).}
}
\value{
NONE
}
\description{
State Variable Data Assimilation: Ensemble Kalman Filter and Generalized ensemble filter. Check out SDA_control function for more details on the control arguments.
}
\author{
Michael Dietze, Ann Raiho and Alexis Helgeson \email{dietze@bu.edu}
}
