% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_parallel.R
\name{sda.qsub.job.submission}
\alias{sda.qsub.job.submission}
\title{sda_assemble}
\usage{
sda.qsub.job.submission(
  batch.folder,
  username,
  outdir = NULL,
  past.job.ids = NULL,
  max.job = 20,
  prefix = "sda.all.forecast.analysis.Rdata",
  resources = list(hour = 24, memory = "4G", ncpu = 28)
)
}
\arguments{
\item{batch.folder}{character: path where the SDA batch jobs stored.}

\item{username}{character: username for the HPC.}

\item{outdir}{character: path where to stored the assembled results. Default is NULL.}

\item{past.job.ids}{vector: vector of the past submitted job ids. Default is NULL.}

\item{prefix}{character: file name to be determined if a job is finished or not.}

\item{resources}{list: computation resources (time, memory, cores) when used to submit jobs.}

\item{max.job:}{numeric: maximum allowance for the number of running jobs on HPC.}
}
\value{
vector or numeric: return -1 when all jobs are finished; 
return vector of submitted job ids when there is any job that is not finished.
}
\description{
This function can help to assemble sda outputs (analysis and forecasts) from each job execution.
}
\author{
Dongchen Zhang.
}
