% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_parallel.R
\name{sda_assemble}
\alias{sda_assemble}
\title{sda_assemble}
\usage{
sda_assemble(batch.folder, outdir)
}
\arguments{
\item{batch.folder}{character: path where the SDA batch jobs stored.}

\item{outdir}{character: path where we want to store the assembled analysis and forecasts.}
}
\description{
This function can help to assemble sda outputs (analysis and forecasts) from each job execution.
}
\author{
Dongchen Zhang.
}
