% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.functions.R
\name{SDA_control}
\alias{SDA_control}
\title{SDA_control}
\usage{
SDA_control(
  trace = TRUE,
  ForewardForecast = FALSE,
  interactivePlot = FALSE,
  TimeseriesPlot = FALSE,
  BiasPlot = FALSE,
  plot.title = NULL,
  facet.plots = FALSE,
  debug = FALSE,
  pause = FALSE,
  Profiling = FALSE,
  OutlierDetection = FALSE
)
}
\arguments{
\item{trace}{Logical if code should print out the progress of SDA .}

\item{ForewardForecast}{Logical if the foreward forecast estimates needs to be read and visulized in time series plots.}

\item{interactivePlot}{Logical if time series plots need to be generated.}

\item{TimeseriesPlot}{Logical if time series plots need to be generated.}

\item{BiasPlot}{Logical if bias plots need to be generated}

\item{plot.title}{Character defining the title of times series plots}

\item{facet.plots}{Logical if the timeseries plots should be faceted based on state variables}

\item{debug}{Logical if the code should stop at some milestones and open an interactive debugging environment}

\item{pause}{Logical if code needs to be puased and wait for further instruction after the analysis step}

\item{Profiling}{Logical if code should keep track of how much time each step took}

\item{OutlierDetection}{Logical if TRUE then a simple method will be used to replace simulations of outside 3IQR with the median of ensembles.}
}
\value{
list of all arguments needed to setup the SDA function
}
\description{
SDA_control
}
