% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_function.R
\name{SDA_downscale_preprocess}
\alias{SDA_downscale_preprocess}
\title{Preprocess Data for Downscaling}
\usage{
SDA_downscale_preprocess(data_path, coords_path, date, carbon_pool)
}
\arguments{
\item{data_path}{Character. File path for .rds containing ensemble data.}

\item{coords_path}{Character. File path for .csv file containing the site
coordinates, with columns named "lon" and "lat".}

\item{date}{Date. If SDA site run, format is yyyy/mm/dd; if NEON, yyyy-mm-dd.
Restricted to years within the file supplied to 'data_path'.}

\item{carbon_pool}{Character. Carbon pool of interest. Name must match the
carbon pool name found within the file supplied to 'data_path'.}
}
\value{
A list containing The read .rds data , The cleaned site coordinates,
  and the preprocessed carbon data.
}
\description{
This function reads and checks the input data, ensuring that the required
date and carbon pool exist, and that the site coordinates are valid.
}
\details{
It ensures that the specified date and carbon pool are present in the input
data. It also checks the validity of the site coordinates and aligns the
number of rows between site coordinates and carbon data.
}
\author{
Sambhav Dixit
}
