% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_matchparam.R
\name{sda_matchparam}
\alias{sda_matchparam}
\title{sda_matchparam}
\usage{
sda_matchparam(settings, ensemble.samples, site.ids, nens)
}
\arguments{
\item{settings}{settings object passed from sda.enkf_MultiSite}

\item{ensemble.samples}{taken from sample.Rdata object}

\item{site.ids}{character object passed from sda.enkf_MultiSite}

\item{nens}{number of ensemble members in model runs, taken from restart$runids}
}
\value{
new.params object used to
}
\description{
sda_matchparam
}
\author{
Alexis Helgeson
}
