% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_OBS_Assembler.R
\name{SDA_OBS_Assembler}
\alias{SDA_OBS_Assembler}
\title{Assembler for preparing obs.mean and obs.cov for the SDA workflow}
\usage{
SDA_OBS_Assembler(settings_dir, var_name, outdir)
}
\arguments{
\item{settings_dir}{the path of settings.xml object.}

\item{var_name}{Variable name, currently support: SMP, AGB, and LAI.}

\item{outdir}{the path to store obs.mean and obs.cov}
}
\value{
list of obs.mean and obs.cov
}
\description{
Assembler for preparing obs.mean and obs.cov for the SDA workflow
}
\examples{
\dontrun{
settings_dir <- "/projectnb/dietzelab/dongchen/All_NEON_SDA/NEON42/IC/pecan.xml"
outdir <- "/projectnb/dietzelab/dongchen/All_NEON_SDA/test_OBS"
var_name <- c("LAI", "AbvGrndWood", "SoilMoistFrac", "TotSoilCarb")
OBS <- SDA_OBS_Assembler(settings_dir, var_name, outdir)
}
}
\author{
Dongchen Zhang
}
