% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_unitnormalization.R
\name{sda_unitnormalization}
\alias{sda_unitnormalization}
\title{SDA unit normalization}
\usage{
sda_unitnormalization(
  X,
  Y = NULL,
  R = NULL,
  Pa = NULL,
  mu.a = NULL,
  method = FALSE
)
}
\arguments{
\item{X}{matrix with state variables created in build_x}

\item{Y}{obs.mean}

\item{R}{obs.cov}

\item{Pa}{analysis mean}

\item{mu.a}{analysis cov}

\item{method}{FALSE = forward, TRUE = reverse}
}
\value{

}
\description{
SDA unit normalization
}
