% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_weights_site.R
\name{sda_weights_site}
\alias{sda_weights_site}
\title{Calculate ensemble weights for each site at time t.}
\usage{
sda_weights_site(FORECAST, ANALYSIS, t, ens)
}
\arguments{
\item{FORECAST}{FORECAST object built within the sda.enkf_MultiSite function.}

\item{ANALYSIS}{ANALYSIS object built within the Analysis_sda_multisite function.}

\item{t}{exact number of t inside the sda.enkf_MultiSite function.}

\item{ens}{number of ensemble members.}
}
\value{
list of weights associated with each ensemble member of each site.
}
\description{
Calculate ensemble weights for each site at time t.
}
\author{
Dongchen Zhang and Hamze Dokoohaki
}
