% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_parallel_downscale.R
\name{stack_covariates_2_df}
\alias{stack_covariates_2_df}
\title{stack_covariates_2_df}
\usage{
stack_covariates_2_df(rast.dir, cores = parallel::detectCores())
}
\arguments{
\item{rast.dir}{character: a character that points to the covariates raster file generated by the `stack_covariates_2_geotiff` function.}

\item{cores}{numeric: how many CPus to be used in the calculation, the default is the total CPU number you have.}
}
\value{
list containing the data frame of covariates for vegetated pixels and the corresponding index of the pixels.
}
\description{
This function helps to build the data frame (pixels by data columns) for only vegetated pixels to improve the efficiency.
Note that the `LC` field using the `MODIS land cover` observations (MCD12Q1.061) must be supplied in the covariates to make this function work.
}
\author{
Dongchen Zhang
}
