% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_parallel_downscale.R
\name{stack_covariates_2_geotiff}
\alias{stack_covariates_2_geotiff}
\title{stack_covariates_2_geotiff}
\usage{
stack_covariates_2_geotiff(
  outdir,
  year,
  base.map.dir,
  cov.tif.file.list,
  normalize = T,
  cores = parallel::detectCores()
)
}
\arguments{
\item{outdir}{character: the output directory where the stacked GeoTIFF file will be generated.}

\item{year}{numeric: the year of when the covariates are stacked.}

\item{base.map.dir}{character: path to the GeoTIFF file within which the extents and CRS will be used to generate the final map.}

\item{cov.tif.file.list}{list: a list contains sub-lists with each including path to the corresponding map and the variables to be extracted (e.g., list(LC = list(dir = "path/to/landcover.tiff", var.name = "LC")).}

\item{normalize}{boolean: decide if we want to normalize each data layer, the default is TRUE.}

\item{cores}{numeric: how many CPus to be used in the calculation, the default is the total CPU number you have.}
}
\value{
path to the exported GeoTIFF file.
}
\description{
This function helps to stack target data layers from various GeoTIFF maps (with different extents, CRS, and resolutions) to a single map.
}
\author{
Dongchen Zhang
}
