% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_downscale.R
\name{subset_ensemble}
\alias{subset_ensemble}
\title{Subset ensemble data for downscaling}
\usage{
subset_ensemble(ensemble_data, site_coords, date, carbon_pool)
}
\arguments{
\item{ensemble_data}{EFI standard tibble or data.frame}

\item{site_coords}{A data.frame with a unique `site_id` column.}

\item{date}{Date. The date for the run, must be a date within `ensemble_data`.}

\item{carbon_pool}{Character. Carbon pool of interest. Name must match a value in `ensemble_data$variable`.}
}
\value{
A data.frame containing `site_id`, `ensemble`, and `prediction` for the specified date and carbon pool.
}
\description{
Subset ensemble data for downscaling
}
\details{
This function subsets ensemble data and ensures that the specified date and
carbon pool are present in the ensemble data.
}
\author{
Sambhav Dixit, David LeBauer
}
