% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEF_Helper.R
\name{tobit_model_censored}
\alias{tobit_model_censored}
\title{tobit_model_censored}
\usage{
tobit_model_censored(settings, X, var.names, mu.f, Pf, t)
}
\arguments{
\item{settings}{}

\item{X}{}

\item{var.names}{}

\item{mu.f}{}

\item{Pf}{}
}
\description{
tobit_model_censored
}
