% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.batch.functions.R
\name{write.configs.fun}
\alias{write.configs.fun}
\title{write.configs.fun}
\usage{
write.configs.fun(folder.path, cores, lib)
}
\arguments{
\item{folder.path}{Character: physical path to which the job file is located.}

\item{cores}{numeric: number of requested CPUs.}

\item{lib}{Character: extra required library.}
}
\description{
This function realizes the parallel `write.configs`.
}
\author{
Dongchen Zhang.
}
