% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remote_helpers.R
\name{alltocs}
\alias{alltocs}
\title{alltocs}
\usage{
alltocs(fname = "tocs.csv")
}
\arguments{
\item{fname}{string path to where the output needs to be saved as a csv file.}
}
\value{
This function writes down a csv file with three columns: 1- message sepecified in the `tic` 2- Total elapsed time and 3- the execution time
}
\description{
This function finds all the tic functions called before and estimates the time elapsed for each one saves/appends it to a csv file.
}
\examples{
 
\dontrun{
 library(tictoc)
 tic("Analysis")
 Sys.sleep(5)
 testfunc()
 tic("Adjustment")
 Sys.sleep(4)
 alltocs("timing.csv")  
}
}
