% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda.R
\name{Analysis.sda}
\alias{Analysis.sda}
\title{Analysis.sda}
\usage{
Analysis.sda(settings, FUN, Forecast = list(Pf = NULL, mu.f = NULL, Q =
  NULL, X = NULL), Observed = list(R = NULL, Y = NULL), ...)
}
\arguments{
\item{settings}{pecan standard settings list.}

\item{FUN}{A Function for performing the analysis step. Two available options are: 1-EnKF and 2-GEF.}

\item{Forecast}{A list containing the forecasts variables including Pf (cov of forecast state variables), mu.f (vector of estimated mean of forecast state variables),
Q (process variance) and X (a dataframe of forcats state variables for different ensemble)}

\item{Observed}{A list containing the observed variables including R (cov of observed state variables) and Y (vector of estimated mean of observed state variables)}

\item{...}{Extra argument sent to the analysis function.}
}
\value{
Returns whatever the FUN is returning. In case of EnKF and GEF, this function returns a list with estimated mean and cov matrix of forecast state variables as well as mean and cov estimated as a result of assimilation/analysis .
}
\description{
This functions uses the FUN to perform the analysis. EnKF function is developed inside the PEcAn.assim.sequential package which can be sent to this function to perform the Ensemble Kalman Filter. 
The other option is GEF function inside the same package allowing to perform Generalized Ensemble kalman Filter.

If you're using an arbitrary function you can use the ... to send any other variables to your desired analysis function.
}
\author{
Michael Dietze \email{dietze@bu.edu}, Ann Raiho and Hamze Dokoohaki
}
