% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_X.R
\name{build_X}
\alias{build_X}
\title{build_X}
\usage{
build_X(
  out.configs,
  settings,
  new.params,
  nens,
  read_restart_times,
  outdir,
  t = 1,
  var.names,
  my.read_restart,
  restart_flag = FALSE
)
}
\arguments{
\item{out.configs}{object created for build_X passed from sda.enkf_MultiSite}

\item{settings}{settings object, passed from sda.enkf_MultiSite}

\item{new.params}{object created from sda_matchparam, passed from sda.enkf_MultiSite}

\item{nens}{number of ensemble members i.e. runs}

\item{read_restart_times}{passed from sda.enkf_MultiSite}

\item{outdir}{location of previous run output folder containing .nc files}

\item{t}{Default t=1, for function to work within time loop}

\item{var.names}{list of state variables taken from settings object}

\item{my.read_restart}{object that points to the model restart function i.e. read_restart.SIPNET}

\item{restart_flag}{flag if it's a restart stage. Default is FALSE.}
}
\value{
X ready to be passed to SDA Analysis code
}
\description{
builds X matrix for SDA
}
\author{
Alexis Helgeson
}
