% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\name{conj_wt_wishart_sampler}
\alias{conj_wt_wishart_sampler}
\title{Weighted conjugate wishart}
\usage{
conj_wt_wishart_sampler(model, mvSaved, target, control)
}
\arguments{
\item{model}{model}

\item{mvSaved}{copied to}

\item{target}{thing being targetted}

\item{control}{unused}
}
\description{
Weighted conjugate wishart
}
