% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\name{dwtmnorm}
\alias{dwtmnorm}
\title{weighted multivariate normal density}
\usage{
dwtmnorm(x, mean, prec, wt, log = 0)
}
\arguments{
\item{x}{random variable}

\item{mean}{mean}

\item{prec}{precision}

\item{wt}{weight}

\item{log}{log}
}
\description{
weighted multivariate normal density
}
