% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_multiSite.R
\name{EnKF.MultiSite}
\alias{EnKF.MultiSite}
\title{EnKF.MultiSite}
\usage{
EnKF.MultiSite(setting, Forecast, Observed, H, extraArg = NULL, ...)
}
\arguments{
\item{Forecast}{A list containing the forecasts variables including Q (process variance) and X (a dataframe of forecasts state variables for different ensemble)}

\item{Observed}{A list containing the observed variables including R (cov of observed state variables) and Y (vector of estimated mean of observed state variables)}

\item{H}{is a mtrix of 1's and 0's specifying which observations go with which variables.}

\item{...}{Extra argument sent to the analysis function.}

\item{settings}{pecan standard settings list.}
}
\value{
It returns a list with estimated mean and cov matrix of forecast state variables as well as mean and cov estimated as a result of assimilation/analysis .
}
\description{
Given the Forecast and Observed this function performs the Ensemble Kalamn Filter.
}
\details{
This function is different than EnKF in terms of how it creates the Pf matrix.
}
\author{
Michael Dietze \email{dietze@bu.edu}, Ann Raiho and Hamze Dokoohaki
}
