% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ensemble_weights.R
\name{get_ensemble_weights}
\alias{get_ensemble_weights}
\title{get_ensemble_weights}
\usage{
get_ensemble_weights(settings, time_do)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{time_do}{Give user specific time so you don't have to have it be annual}
}
\value{
NONE
}
\description{
Creates file of ensemble weights in format needed for SDA
}
\author{
Ann Raiho \email{ann.raiho@gmail.com}
}
