% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_plotting.R
\name{interactive.plotting.sda}
\alias{interactive.plotting.sda}
\alias{postana.timeser.plotting.sda}
\alias{postana.bias.plotting.sda}
\alias{postana.bias.plotting.sda.corr}
\alias{post.analysis.ggplot}
\alias{post.analysis.ggplot.violin}
\alias{post.analysis.multisite.ggplot}
\alias{SDA_timeseries_plot}
\title{Internal functions for plotting SDA outputs. Interactive, post analysis time-series and bias plots in base plotting system and ggplot}
\usage{
interactive.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.timeser.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.bias.plotting.sda(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS
)

postana.bias.plotting.sda.corr(t, obs.times, X, aqq, bqq)

post.analysis.ggplot(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS,
  plot.title = NULL
)

post.analysis.ggplot.violin(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  obs,
  X,
  FORECAST,
  ANALYSIS,
  plot.title = NULL
)

post.analysis.multisite.ggplot(
  settings,
  t,
  obs.times,
  obs.mean,
  obs.cov,
  FORECAST,
  ANALYSIS,
  plot.title = NULL,
  facetg = FALSE,
  readsFF = NULL,
  Add_Map = FALSE
)

SDA_timeseries_plot(
  ANALYSIS,
  FORECAST,
  obs.mean = NULL,
  obs.cov = NULL,
  outdir,
  pft.path = NULL,
  by = "site",
  types = c("FORECAST", "ANALYSIS", "OBS"),
  CI = c(0.025, 0.975),
  unit = list(AbvGrndWood = "Mg/ha", LAI = "m2/m2", SoilMoistFrac = "", TotSoilCarb =
    "kg/m2"),
  style = list(general_color = c(FORECAST = "blue", ANALYSIS = "red", OBS = "black"),
    fill_color = c(FORECAST = "yellow", ANALYSIS = "green", OBS = "grey"), title_color =
    "red"),
  PDF_w = 20,
  PDF_h = 16,
  t.inds = NULL
)
}
\arguments{
\item{settings}{pecan standard settings list.}

\item{t}{current time - int number giving the position of the current time in obs.time.}

\item{obs.times}{vector of dates of measurements}

\item{obs.mean}{obs.mean}

\item{obs.cov}{obs.cov}

\item{obs}{list containing the mean and cov object}

\item{X}{dataframe of state variables for each ensemble}

\item{FORECAST}{Forecast object from the sda.output.Rdata.}

\item{ANALYSIS}{Analysis object from the sda.output.Rdata.}

\item{plot.title}{character giving the title for post visualization ggplots}

\item{Add_Map}{Bool variable decide if we want to export the GIS map of Ecoregion.}

\item{outdir}{physical path where the pdf will be stored.}

\item{pft.path}{Physical path of pft.csv file to allow by = pft option.}

\item{by}{arrange figures by var, pft, or site.}

\item{types}{data types that shown in the figure.}

\item{CI}{range of confidence interval.}

\item{unit}{list of unit used for y axis label.}

\item{style}{color option.}

\item{PDF_w}{width of exported PDF file, passed on to `base::pdf()`.}

\item{PDF_h}{height of exported PDF file, passed on to `base::pdf()`.}

\item{t.inds}{index of period that will be plotted.}
}
\description{
Internal functions for plotting SDA outputs. Interactive, post analysis time-series and bias plots in base plotting system and ggplot
}
\author{
Dongchen Zhang
}
