% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\name{load_nimble}
\alias{load_nimble}
\title{load_nimble}
\usage{
load_nimble()
}
\description{
This functions is used internally to register load nimble library and required functions.
}
\author{
Ann Raiho
}
