% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\name{load_nimble}
\alias{load_nimble}
\alias{y_star_create}
\title{load_nimble}
\usage{
y_star_create(X)
}
\arguments{
\item{X}{state var}
}
\description{
This functions is internally used to register a series of nimble functions inside GEF analysis function.
}
\author{
Ann Raiho, Hamze Dokoohaki
}
