% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remote_helpers.R
\name{Obs.data.prepare.MultiSite}
\alias{Obs.data.prepare.MultiSite}
\title{Obs.data.prepare.MultiSite}
\usage{
Obs.data.prepare.MultiSite(obs.path, site.ids)
}
\arguments{
\item{obs.path}{Path to the obs data which is expected to be an .Rdata.}

\item{site.ids}{a character vector of site ids which need to be extracted.}
}
\value{
a list of observed mean and cov as the SDA expected it to be.
}
\description{
Obs.data.prepare.MultiSite
}
