% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.batch.functions.R
\name{qsub_analysis}
\alias{qsub_analysis}
\title{qsub_analysis}
\usage{
qsub_analysis(folder.path, cores)
}
\arguments{
\item{folder.path}{character: path where the `block.Rdata` file is stored.}

\item{cores}{numeric: number of cpus used for parallel computaion. Default is NULL.}
}
\description{
This function can help to execute `foreach` parallel MCMC sampling given generated MCMC configuration lists.
}
\author{
Dongchen Zhang.
}
