% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R
\name{qsub_analysis_submission}
\alias{qsub_analysis_submission}
\title{qsub_analysis_submission}
\usage{
qsub_analysis_submission(block.list, outdir, job.per.folder = 200)
}
\arguments{
\item{block.list}{list: MCMC configuration lists for the block SDA analysis.}

\item{outdir}{character: SDA output path.}

\item{job.per.folder}{numeric: number of jobs per folder.}
}
\description{
This function provides means to split large SDA analysis (MCMC) runs into separate `qsub` jobs.
Including job creation, submission, and assemble.
}
