% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_parallel.R
\name{qsub_sda}
\alias{qsub_sda}
\title{qsub_sda}
\usage{
qsub_sda(
  settings,
  obs.mean,
  obs.cov,
  Q,
  pre_enkf_params,
  ensemble.samples,
  covariates_df,
  outdir,
  control,
  block.index = NULL
)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{obs.mean}{Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation means for each state variables of each site for each time point.}

\item{obs.cov}{Lists of date times named by time points, which contains lists of sites named by site ids, which contains observation covariances for all state variables of each site for each time point.}

\item{Q}{Process covariance matrix given if there is no data to estimate it.}

\item{pre_enkf_params}{Used for passing pre-existing time-series of process error into the current SDA runs to ignore the impact by the differences between process errors.}

\item{ensemble.samples}{Pass ensemble.samples from outside to avoid GitHub check issues.}

\item{outdir}{Physical path to the folder where the SDA outputs will be stored. 
The default is NULL, where we will be using outdir from the settings object.}

\item{control}{List of flags controlling the behavior of the SDA. 
`TimeseriesPlot` for post analysis examination; 
`OutlierDetection` decide if we want to execute the outlier detection each time after the model forecasting;
`send_email` contains lists for sending email to report the SDA progress;
`keepNC` decide if we want to keep the NetCDF files inside the out directory;
`forceRun` decide if we want to proceed the Bayesian MCMC sampling without observations;
`MCMC.args` include lists for controling the MCMC sampling process (iteration, nchains, burnin, and nthin.).}

\item{block.index}{list of site ids for each block, default is NULL. This is used when the localization turns on.
Please keep using the default value because the localization feature is still in development.}
}
\value{
NONE
}
\description{
This function provides means to split large SDA runs into separate `qsub` jobs.
Including job creation, submission, and assemble.
}
\author{
Dongchen Zhang
}
