% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.batch.functions.R
\name{rm.files}
\alias{rm.files}
\title{rm.files}
\usage{
rm.files(folder.path, cores, only.nc = T)
}
\arguments{
\item{folder.path}{Character: physical path to which the job file is located.}

\item{cores}{numeric: number of requested CPUs.}
}
\description{
This function realizes the parallel `unlink` function.
}
\author{
Dongchen Zhang.
}
