% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample.parameters}
\alias{sample.parameters}
\title{sample parameters}
\usage{
sample.parameters(ne, settings, con)
}
\arguments{
\item{ne}{number of ensemble members}

\item{settings}{PEcAn settings object}

\item{con}{PEcAn database connection}
}
\value{
data frame of sampled parameters from the posterior distribution
}
\description{
sample parameters
}
\author{
Michael Dietze \email{dietze@bu.edu}
}

