% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.functions.R
\name{SDA.arguments}
\alias{SDA.arguments}
\title{SDA.arguments}
\usage{
SDA.arguments(trace = TRUE, ForewardForecast = FALSE,
  interactivePlot = FALSE, TimeseriesPlot = FALSE, BiasPlot = FALSE,
  plot.title = NULL, facet.plots = FALSE, debug = FALSE,
  pause = FALSE, Profiling = FALSE)
}
\arguments{
\item{trace}{Logical if code should print out the progress of SDA .}

\item{ForewardForecast}{Logical if the foreward forecast estimates needs to be read and visulized in time series plots.}

\item{interactivePlot}{Logical if time series plots need to be generated.}

\item{TimeseriesPlot}{Logical if time series plots need to be generated.}

\item{BiasPlot}{Logical if bias plots need to be generated}

\item{plot.title}{Character defining the title of times series plots}

\item{facet.plots}{Logical if the timeseries plots should be faceted based on state variables}

\item{debug}{Logical if the code should stop at some milestones and open an interactive debugging environment}

\item{pause}{Logical if code needs to be puased and wait for further instruction after the analysis step}

\item{Profiling}{Logical if code should keep track of how much time each step took}
}
\value{
list of all arguments needed to setup the SDA function
}
\description{
SDA.arguments
}
