% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda.enkf_refactored.R
\name{sda.enkf}
\alias{sda.enkf}
\title{State Variable Data Assimilation: Ensemble Kalman Filter and Generalized ensemble filter}
\usage{
sda.enkf(
  settings,
  obs.mean,
  obs.cov,
  Q = NULL,
  restart = NULL,
  control = list(trace = TRUE, interactivePlot = TRUE, TimeseriesPlot = TRUE, BiasPlot =
    FALSE, plot.title = NULL, debug = FALSE, pause = FALSE),
  ...
)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{obs.mean}{List of dataframe of observation means, named with
observation datetime.}

\item{obs.cov}{List of covariance matrices of state variables , named with
observation datetime.}

\item{Q}{Process covariance matrix given if there is no data to
estimate it.}

\item{restart}{Used for iterative updating previous forecasts. When the
restart is TRUE it read the object in SDA folder written from previous
SDA.}

\item{control}{List of flags controlling the behaviour of the SDA. trace
for reporting back the SDA outcomes, interactivePlot for plotting the
outcomes after each step, TimeseriesPlot for post analysis examination,
BiasPlot for plotting the correlation between state variables, plot.title
is the title of post analysis plots and debug mode allows for pausing the
code and examining the variables inside the function.}

\item{...}{Additional arguments, currently ignored}
}
\value{
NONE
}
\description{
Restart mode:  Basic idea is that during a restart (primary case
 envisioned as an iterative forecast), a new workflow folder is created and
 the previous forecast for the start_time is copied over. During restart the
 initial run before the loop is skipped, with the info being populated from
 the previous run. The function then dives right into the first Analysis,
 then continues on like normal.
}
\author{
Michael Dietze and Ann Raiho \email{dietze@bu.edu}
}
