% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sda.enkf}
\alias{sda.enkf}
\title{sda.enkf}
\usage{
sda.enkf(settings, IC, prior, obs)
}
\arguments{
\item{settings}{PEcAn settings object}

\item{IC}{data.frame of initial condition sample (nens X nstate)}

\item{prior}{data.frame of model parameter sample (nense X nstate)}

\item{obs}{data.frame of observations with columns: mean, sd}
}
\value{
NONE
}
\description{
State Variable Data Assimilation: Ensemble Kalman Filter
}
\author{
Michael Dietze \email{dietze@bu.edu}
}

