% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_function_hrly.R
\name{SDA_downscale_hrly}
\alias{SDA_downscale_hrly}
\title{SDA　Downscale Function for Hourly Data}
\usage{
SDA_downscale_hrly(nc_file, coords, yyyy, covariates)
}
\arguments{
\item{nc_file}{In quotes, file path for .nc containing ensemble data.}

\item{coords}{In quotes, file path for .csv file containing the site coordinates, columns named "lon" and "lat".}

\item{yyyy}{In string, format is yyyy(year of interest)}

\item{covariates}{SpatRaster stack, used as predictors in randomForest. Layers within stack should be named. Recommended that this stack be generated using 'covariates' instructions in assim.sequential/inst folder}
}
\value{
It returns the `downscale_output` list containing lists for the training and testing data sets, models, and predicted maps for each ensemble member.
}
\description{
This function uses the randomForest model to downscale forecast data (hourly) to unmodeled locations using covariates and site locations
}
\author{
Harunobu Ishii
}
