% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_function.R
\name{SDA_downscale_metrics}
\alias{SDA_downscale_metrics}
\title{Calculate Metrics for Downscaling Results}
\usage{
SDA_downscale_metrics(downscale_output, carbon_pool)
}
\arguments{
\item{downscale_output}{List. Output from the SDA_downscale function, containing data, models, maps, and predictions for each ensemble.}

\item{carbon_pool}{Character. Name of the carbon pool used in the downscaling process.}
}
\value{
A list of metrics for each ensemble, where each element contains MAE , MSE ,R_squared ,actual values from testing data and predicted values for the testing data
}
\description{
This function takes the output from the SDA_downscale function and computes various performance metrics for each ensemble. It provides a way to evaluate the accuracy of the downscaling results without modifying the main downscaling function.
}
\details{
This function calculates performance metrics for the downscaling results. It computes Mean Squared Error (MSE), Mean Absolute Error (MAE), and R-squared for each ensemble. The function uses the actual values from the testing data and the predictions generated during the downscaling process.
}
\author{
Sambhav Dixit
}
