% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_function.R
\name{SDA_downscale_preprocess}
\alias{SDA_downscale_preprocess}
\title{Preprocess Data for Downscaling}
\usage{
SDA_downscale_preprocess(ensemble_data, site_coords, date, carbon_pool)
}
\arguments{
\item{ensemble_data}{Character or list. Either a file path to an `.rds` file containing ensemble data, or an in-memory list where each element is a data.frame indexed by date.}

\item{date}{Date. If SDA site run, format is yyyy/mm/dd; if NEON, yyyy-mm-dd. Restricted to years within the file or object supplied to 'ensemble_data'.}

\item{carbon_pool}{Character. Carbon pool of interest. Name must match the carbon pool name found within the file or object supplied to 'ensemble_data'.}

\item{site_coordinates}{Character, data.frame, or sf object. Either:
- A file path to a `.csv` containing site coordinates with columns `"id"`, `"lat"`, and `"lon"`.
- A `data.frame` with the same structure.
- An `sf` object with point geometries.}
}
\value{
A list containing The read .rds data , The cleaned site coordinates, and the preprocessed carbon data.
}
\description{
This function reads and checks the input data, ensuring that the required date and carbon pool exist, and that the site coordinates are valid.
}
\details{
This function ensures that the specified date and carbon pool are present in the input data. It also checks the validity of the site coordinates and aligns the number of rows between site coordinates and carbon data.
}
\author{
Sambhav Dixit

David LeBauer
}
