% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_function.R, R/ensemble_downscale.R
\name{SDA_downscale_preprocess}
\alias{SDA_downscale_preprocess}
\alias{subset_ensemble}
\title{Preprocess Data for Downscaling}
\usage{
SDA_downscale_preprocess(ensemble_data, site_coords, date, carbon_pool)

subset_ensemble(ensemble_data, site_coords, date, carbon_pool)
}
\arguments{
\item{ensemble_data}{EFI standard tibble or data.frame}

\item{site_coords}{data.frame with unique site id}

\item{date}{Date. The date for the run, must be a date within `ensemble_data`.}

\item{carbon_pool}{Character. Carbon pool of interest. Name must match the carbon pool name in ensemble_data. 
found within the file or object supplied to 'ensemble_data'.}

\item{site_coordinates}{Character, data.frame, or sf object. Either:
- A file path to a `.csv` containing site coordinates with columns `"id"`, `"lat"`, and `"lon"`.
- A `data.frame` with the same structure.
- An `sf` object with point geometries.}
}
\value{
A list containing The read .rds data , The cleaned site coordinates, and the preprocessed carbon data.

A list containing the cleaned site coordinates and the ensemble carbon output for the 
specified date and carbon pool.
}
\description{
This function reads and checks the input data, ensuring that the required date and carbon pool exist, and that the site coordinates are valid.
}
\details{
This function ensures that the specified date and carbon pool are present in the input data. It also checks the validity of the site coordinates and aligns the number of rows between site coordinates and carbon data.

This function subsets ensemble data and ensures that the specified date and 
carbon pool are present in the ensemble data.
}
\author{
Sambhav Dixit

David LeBauer

Sambhav Dixit, David LeBauer
}
