% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_OBS_Assembler.R
\name{SDA_OBS_Assembler}
\alias{SDA_OBS_Assembler}
\title{Assembler for preparing obs.mean and obs.cov for the SDA workflow}
\usage{
SDA_OBS_Assembler(settings_dir, Var, OutDir, Obs_Prep = NULL)
}
\arguments{
\item{settings_dir}{the path of settings.xml object.}

\item{Var}{Variable name, currently support: SMP, AGB, and LAI.}

\item{OutDir}{the path to store obs.mean and obs.cov}

\item{Obs_Prep}{if your settings object doesn't contain Obs_Prep, you can import it separately (details see L17-18).}
}
\value{
list of obs.mean and obs.cov
}
\description{
Assembler for preparing obs.mean and obs.cov for the SDA workflow
}
\examples{
settings_dir <- "/projectnb/dietzelab/dongchen/All_NEON_SDA/NEON42/SDA/pecan.xml"
OutDir <- "/projectnb/dietzelab/dongchen/All_NEON_SDA/test_OBS"
Var <- c("SMP", "LAI", "AGB")
OBS <- PEcAnAssimSequential::SDA_OBS_Assembler(settings_dir, Var, OutDir)
}
\author{
Dongchen Zhang
}
