% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDA_OBS_Assembler.R
\name{SDA_OBS_Assembler}
\alias{SDA_OBS_Assembler}
\title{Assembler for preparing obs.mean and obs.cov for the SDA workflow}
\usage{
SDA_OBS_Assembler(settings)
}
\arguments{
\item{settings}{the settings object followed by PEcAn.settings format.}
}
\value{
list of obs.mean and obs.cov
}
\description{
Assembler for preparing obs.mean and obs.cov for the SDA workflow
}
\examples{
\dontrun{
settings_dir <- "/projectnb/dietzelab/dongchen/All_NEON_SDA/NEON42/IC/pecan.xml"
settings <- PEcAn.settings::read.settings(settings_dir)
OBS <- SDA_OBS_Assembler(settings)
}

}
\author{
Dongchen Zhang
}
