% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Localization.R
\name{simple.local}
\alias{simple.local}
\title{simple.local}
\usage{
simple.local(Pf, distance.mat, scalef = 2)
}
\arguments{
\item{Pf}{Forecast error coveriance matrix}

\item{distance.mat}{is matrix of distances between sites.}

\item{scalef}{scalef is a numeric value that requires tunning and it controls the shape of the corrolation function}
}
\value{
It returns a localized covariance matrix by taking a Schur product between Pf and a corrolation function
}
\description{
Adopted from Data assimilation for spatio-temporal processes - p250 - Sebastian Reich
}
\author{
Hamze Dokoohaki
}
