% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nimble_codes.R
\docType{data}
\name{tobit2space.model}
\alias{tobit2space.model}
\title{Fit tobit prior to ensemble members}
\format{
TBD
}
\usage{
tobit2space.model
}
\description{
Fit tobit prior to ensemble members
}
\keyword{datasets}
