% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEF_Helper.R
\name{tobit_model_censored}
\alias{tobit_model_censored}
\title{tobit_model_censored}
\usage{
tobit_model_censored(settings, X, var.names, mu.f, Pf, t)
}
\arguments{
\item{settings}{(list) pecan standard settings list.}

\item{X}{(numeric) A matrix contains ensemble forecasts (ensembles * variables).}

\item{var.names}{(character) variable names.}

\item{mu.f}{(numeric) forecast mean values.}

\item{Pf}{(numeric) forecast covariance matrix.}

\item{t}{(numeric) timestep. If t=1, initial values are imputed for zero values in mu.f}
}
\value{
list with updated mu.f, pf, X, and indication of which y values are censored
}
\description{
tobit_model_censored
}
