% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_sda_block.R
\name{update_q}
\alias{update_q}
\title{update_q}
\usage{
update_q(
  block.list.all,
  t,
  nt,
  aqq.Init = NULL,
  bqq.Init = NULL,
  MCMC_dat = NULL,
  block.list.all.pre = NULL
)
}
\arguments{
\item{block.list.all}{each block within the `block.list` lists.}

\item{t}{time point.}

\item{nt}{total length of time steps.}

\item{aqq.Init}{the initial values of aqq, the default is NULL.}

\item{bqq.Init}{the initial values of bqq, the default is NULL.}

\item{MCMC_dat}{data frame of MCMC samples, the default it NULL.}

\item{block.list.all.pre}{pre-existed block.list.all object for passing the aqq and bqq to the current SDA run, the default is NULL.}
}
\value{
It returns the `block.list.all` object with initialized/updated Q filled in.
}
\description{
update_q
}
\author{
Dongchen Zhang
}
